fix(menu-scroll): attempt to fix horizontal scroll issue on burger menu appearance
This commit attempts to fix the issue of horizontal scrolling on the
website when the burger menu appears. The changes include the addition
of a new container in the HTML that wraps all the content on the site.
This container is given the CSS property 'overflow: hidden' to restrict
any overflow of content, thereby addressing the horizontal scroll issue.